python-service/requirements.txt
# Gemini AI for table extraction and OCR
google-generativeai>=0.8.0

# Excel Generation
openpyxl>=3.1.0

# Web Service
Flask>=3.0.0
Flask-CORS>=4.0.0
gunicorn>=21.0.0

# Image Processing
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
